Feat/order flow

+++ src/main/java/com/CapStone/blinkitservice/cart/CartService.java
@@ -127,9 +127,9 @@ public class CartService {
         for(CartItemEntity cartItem:userCartItems){
             ProductEntity productEntity=cartItem.getProductEntity();
 
-            Double discountApplied = 0.0;
+            Float discountApplied = 0.0f;
             if(productEntity.getDiscount() != null){
-                discountApplied = (productEntity.getPrice()*(productEntity.getDiscount()/100.0));
+                discountApplied = productEntity.getPrice()*(productEntity.getDiscount()/100);
             }
 
             UpdateCartProductResponse productResponse = UpdateCartProductResponse.builder()
@@ -144,8 +144,8 @@ public class CartService {
                     .isAvailable(productEntity.isAvailable())
                     .build();
 
-            totalWithoutDiscount += productEntity.getPrice().floatValue()*cartItem.getQuantity();
-            grandTotal += productResponse.getDiscountedPrice().floatValue()*cartItem.getQuantity();
+            totalWithoutDiscount += productEntity.getPrice()*cartItem.getQuantity();
+            grandTotal += productResponse.getDiscountedPrice()*cartItem.getQuantity();
             uniqueQuantity += 1;
             totalQuantity += cartItem.getQuantity();
             productResponses.add(productResponse);
